third_party/WebKit/LayoutTests/fast/dom/setPrimitiveValue-exceptions-expected.txt

Test exceptions thrown by the CSSPrimitiveValue APIs. Primitive values are currently immutable (see https://bugs.webkit.org/show_bug.cgi?id=31223)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS left.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_DIMENSION) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_PX) is 10
PASS left.setFloatValue(CSSPrimitiveValue.CSS_NUMBER, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_DIMENSION, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_PX, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_UNKNOWN, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_STRING, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.getFloatValue(CSSPrimitiveValue.CSS_UNKNOWN) threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getFloatValue(CSSPrimitiveValue.CSS_STRING) threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getStringValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getCounterValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getRectValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getRGBColorValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_DIMENSION) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_PX) is 10
PASS fontFamily.getStringValue() is "Times"
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_STRING, 'Hi there!') threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_ATTR, "G'day!") threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_UNKNOWN, 'Hi there!') threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_DIMENSION, "G'day!") threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_COUNTER, 'Hello, world!') threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS fontFamily.getFloatValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS fontFamily.getCounterValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS fontFamily.getRectValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS fontFamily.getRGBColorValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS fontFamily.getStringValue() is "Times"
PASS left.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_DIMENSION) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_PX) is 10
PASS left.setFloatValue(CSSPrimitiveValue.CSS_NUMBER, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_DIMENSION, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_PX, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_UNKNOWN, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.setFloatValue(CSSPrimitiveValue.CSS_STRING, 25) threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS left.getFloatValue(CSSPrimitiveValue.CSS_UNKNOWN) threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getFloatValue(CSSPrimitiveValue.CSS_STRING) threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getStringValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getCounterValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getRectValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getRGBColorValue() threw exception InvalidAccessError: A parameter or an operation was not supported by the underlying object..
PASS left.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_DIMENSION) is 10
PASS left.getFloatValue(CSSPrimitiveValue.CSS_PX) is 10
FAIL fontFamily.getStringValue() should be Times. Threw exception TypeError: Cannot call method 'getStringValue' of undefined
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_STRING, 'Hi there!') threw exception TypeError: Cannot call method 'setStringValue' of undefined.
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_ATTR, "G'day!") threw exception TypeError: Cannot call method 'setStringValue' of undefined.
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_UNKNOWN, 'Hi there!') threw exception TypeError: Cannot call method 'setStringValue' of undefined.
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_DIMENSION, "G'day!") threw exception TypeError: Cannot call method 'setStringValue' of undefined.
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_COUNTER, 'Hello, world!') threw exception TypeError: Cannot call method 'setStringValue' of undefined.
PASS fontFamily.getFloatValue() threw exception TypeError: Cannot call method 'getFloatValue' of undefined.
PASS fontFamily.getCounterValue() threw exception TypeError: Cannot call method 'getCounterValue' of undefined.
PASS fontFamily.getRectValue() threw exception TypeError: Cannot call method 'getRectValue' of undefined.
PASS fontFamily.getRGBColorValue() threw exception TypeError: Cannot call method 'getRGBColorValue' of undefined.
FAIL fontFamily.getStringValue() should be Times. Threw exception TypeError: Cannot call method 'getStringValue' of undefined
PASS successfullyParsed is true

TEST COMPLETE
